Extract widgets and update song filter page
