+++ lib/pages/search/search_song_filter_page.dart
@@ -5,8 +5,11 @@ import 'package:vocadb/blocs/search_song_filter_bloc.dart';
 import 'package:vocadb/constants.dart';
 import 'package:vocadb/models/artist_model.dart';
 import 'package:vocadb/models/tag_model.dart';
+import 'package:vocadb/pages/search/artist_stream_filter.dart';
 import 'package:vocadb/pages/search/search_artist_page.dart';
 import 'package:vocadb/pages/search/search_tag_page.dart';
+import 'package:vocadb/pages/search/tag_stream_filter.dart';
+import 'package:vocadb/pages/setting/single_choice_page.dart';
 import 'package:vocadb/widgets/space_divider.dart';
 
 class SearchSongFilterPage extends StatelessWidget {
@@ -38,74 +41,29 @@ class SearchSongFilterPage extends StatelessWidget {
           padding: EdgeInsets.all(8.0),
           child: ListView(
             children: <Widget>[
-              Title(
-                text: FlutterI18n.translate(context, 'label.tags'),
-              ),
-              SpaceDivider(),
-              StreamBuilder(
-                stream: bloc.tags$,
-                builder: (context, snapshot) {
-                  List<TagModel> tags = (snapshot.hasData)
-                      ? (snapshot.data as Map<int, TagModel>).values.toList()
-                      : [];
-
-                  return TagFilters(
-                    tags: tags,
-                    onBrowseTags: () {
-                      browseTags(context);
-                    },
-                    onDeleteTag: (TagModel t) {
-                      bloc.removeTag(t.id);
-                    },
-                  );
+              TagStreamFilters(
+                tags$: bloc.tags$,
+                onDeleteTag: (TagModel tag) {
+                  bloc.removeTag(tag.id);
                 },
+                onBrowseTags: () => browseTags(context),
               ),
-              SpaceDivider(),
-              Title(
-                text: FlutterI18n.translate(context, 'label.songType'),
+              Divider(),
+              SongTypeSelector(
+                value: bloc.songType,
+                onSelected: bloc.updateSongType,
               ),
-              StreamBuilder(
-                stream: bloc.songType$,
-                builder: (context, snapshot) {
-                  return SongTypeDropDown(
-                    value: snapshot.data,
-                    onChanged: bloc.updateSongType,
-                  );
+              SongSortSelector(
+                value: bloc.sort,
+                onSelected: bloc.updateSort,
+              ),
+              Divider(),
+              ArtistStreamFilters(
+                artists$: bloc.artists$,
+                onDeleteArtist: (ArtistModel artist) {
+                  bloc.removeArtist(artist.id);
                 },
-              ),
-              SpaceDivider(),
-              Title(
-                text: FlutterI18n.translate(context, 'label.sort'),
-              ),
-              StreamBuilder(
-                stream: bloc.sort$,
-                builder: (context, snapshot) {
-                  return SongSortDropDown(
-                    value: snapshot.data ?? 'Name',
-                    onChanged: bloc.updateSort,
-                  );
-                },
-              ),
-              SpaceDivider(),
-              Title(
-                text: FlutterI18n.translate(context, 'label.artists'),
-              ),
-              SpaceDivider(),
-              StreamBuilder(
-                stream: bloc.artists$,
-                builder: (context, snapshot) {
-                  List<ArtistModel> artists = (snapshot.hasData)
-                      ? (snapshot.data as Map<int, ArtistModel>).values.toList()
-                      : [];
-
-                  return ArtistFilters(
-                    artists: artists,
-                    onRemove: bloc.removeArtist,
-                    onBrowseArtists: () {
-                      browseArtists(context);
-                    },
-                  );
-                },
+                onBrowseArtists: () => browseArtists(context),
               ),
             ],
           ),
@@ -113,153 +71,46 @@ class SearchSongFilterPage extends StatelessWidget {
   }
 }
 
-class ArtistFilters extends StatelessWidget {
-  final List<ArtistModel> artists;
-  final Function onBrowseArtists;
-  final Function onRemove;
+class SongTypeSelector extends StatelessWidget {
+  final String value;
+  final Function onSelected;
 
-  const ArtistFilters(
-      {Key key, this.artists, this.onBrowseArtists, this.onRemove})
+  const SongTypeSelector({Key key, this.value, this.onSelected})
       : super(key: key);
 
-  List<Widget> buildChildren(BuildContext context) {
-    List<Widget> children = [];
+  List<ChoiceOption> options(BuildContext context) {
+    List<ChoiceOption> options = [];
+    options.add(ChoiceOption(
+        FlutterI18n.translate(context, 'label.notSpecified'), null));
+    options.addAll(constSongTypes
+        .map((v) =>
+            ChoiceOption(FlutterI18n.translate(context, 'songType.$v'), v))
+        .toList());
 
-    if (artists != null && artists.length > 0) {
-      children.addAll(artists
-          .map((t) => ListTile(
-                onTap: () {},
-                trailing: IconButton(
-                  icon: Icon(Icons.close),
-                  onPressed: () {
-                    this.onRemove(t.id);
-                  },
-                ),
-                leading: buildLeading(t.imageUrl),
-                title: Text(t.name),
-                subtitle: Text(t.artistType),
-              ))
-          .toList());
-    }
-
-    children.add(ListTile(
-      onTap: this.onBrowseArtists,
-      leading: Icon(Icons.add),
-      title: Text(FlutterI18n.translate(context, 'label.add')),
-    ));
-
-    return children;
-  }
-
-  Widget buildLeading(String imageUrl) {
-    return SizedBox(
-      width: 50,
-      height: 50,
-      child: ClipOval(
-          child: Container(
-        color: Colors.white,
-        child: (imageUrl == null)
-            ? Placeholder()
-            : CachedNetworkImage(
-                imageUrl: imageUrl,
-                placeholder: (context, url) => Container(color: Colors.grey),
-                errorWidget: (context, url, error) => new Icon(Icons.error),
-              ),
-      )),
-    );
+    return options;
   }
 
   @override
   Widget build(BuildContext context) {
     return Column(
-      children: buildChildren(context),
+      children: <Widget>[
+        ListTile(
+          onTap: () {
+            SingleChoicePage.navigate(
+              context,
+              title: FlutterI18n.translate(context, 'label.songType'),
+              options: options(context),
+              value: value,
+              onSelected: onSelected,
+            );
+          },
+          title: Text(FlutterI18n.translate(context, 'label.songType')),
+          subtitle: Text((value == null)
+              ? FlutterI18n.translate(context, 'label.notSpecified')
+              : FlutterI18n.translate(context, 'songType.$value')),
+        ),
+      ],
     );
-  }
-}
-
-class TagFilters extends StatelessWidget {
-  final Function onBrowseTags;
-  final List<TagModel> tags;
-  final Function onDeleteTag;
-
-  const TagFilters({Key key, this.onBrowseTags, this.tags, this.onDeleteTag})
-      : super(key: key);
-
-  List<Widget> buildChildren(BuildContext context) {
-    List<Widget> children = [];
-
-    if (tags != null && tags.length > 0) {
-      children.addAll(tags
-          .map((t) => Chip(
-                label: Text(t.name),
-                onDeleted: () {
-                  this.onDeleteTag(t);
-                },
-                deleteIcon: Icon(Icons.close),
-              ))
-          .toList());
-    }
-
-    children.add(InputChip(
-      label: Wrap(
-        crossAxisAlignment: WrapCrossAlignment.center,
-        children: <Widget>[
-          Icon(Icons.add),
-          Text(FlutterI18n.translate(context, 'label.add'))
-        ],
-      ),
-      onPressed: this.onBrowseTags,
-    ));
-
-    return children;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Wrap(
-      spacing: 8.0,
-      children: buildChildren(context),
-    );
-  }
-}
-
-class SongTypeDropDown extends StatelessWidget {
-
-  final Function onChanged;
-  final String value;
-
-  const SongTypeDropDown({Key key, this.onChanged, this.value})
-      : super(key: key);
-
-  List<DropdownMenuItem<String>> dropDownItems(BuildContext context) {
-    List<DropdownMenuItem<String>> items = [];
-    items.add(defaultItem(context));
-    items.addAll(constSongTypes.map((v) => createItem(context, v)).toList());
-    
-    return items;
-  }
-
-  DropdownMenuItem<String> defaultItem(BuildContext context) {
-    return DropdownMenuItem<String>(
-      value: null,
-      child: Text(FlutterI18n.translate(context, 'label.notSpecified')),
-    );
-  }
-
-  DropdownMenuItem<String> createItem(BuildContext context, String value) {
-    return DropdownMenuItem<String>(
-      value: value,
-      child: Text(FlutterI18n.translate(context, 'songType.$value')),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return DropdownButton(
-        value: this.value,
-        underline: Container(),
-        items: dropDownItems(context),
-        onChanged: this.onChanged);
   }
 }
 
@@ -291,6 +142,55 @@ class SongSortDropDown extends StatelessWidget {
                 ))
             .toList(),
         onChanged: this.onChanged);
+  }
+}
+
+class SongSortSelector extends StatelessWidget {
+  final String value;
+  final Function onSelected;
+  final values = const [
+    'Name',
+    'AdditionDate',
+    'PublishDate',
+    'FavoritedTimes',
+    'RatingScore'
+  ];
+
+  const SongSortSelector({Key key, this.value, this.onSelected})
+      : super(key: key);
+
+  List<ChoiceOption> options(BuildContext context) {
+    List<ChoiceOption> options = [];
+    options.add(ChoiceOption(
+        FlutterI18n.translate(context, 'label.notSpecified'), null));
+    options.addAll(values
+        .map((v) => ChoiceOption(FlutterI18n.translate(context, 'sort.$v'), v))
+        .toList());
+
+    return options;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: <Widget>[
+        ListTile(
+          onTap: () {
+            SingleChoicePage.navigate(
+              context,
+              title: FlutterI18n.translate(context, 'label.sort'),
+              options: options(context),
+              value: value,
+              onSelected: onSelected,
+            );
+          },
+          title: Text(FlutterI18n.translate(context, 'label.sort')),
+          subtitle: Text((value == null)
+              ? FlutterI18n.translate(context, 'label.notSpecified')
+              : FlutterI18n.translate(context, 'sort.$value')),
+        ),
+      ],
+    );
   }
 }
